:nail_care: Refactor sidebaar menu items
