--- tutorials/postboard/client/src/hooks/types.tsx
@@ -1,7 +1,0 @@
-import { FunctionComponent, SVGProps } from 'react';
-
-export type MenuItem = {
-  iconComponent: FunctionComponent<SVGProps<SVGSVGElement>>;
-  label: string;
-  route: string;
-};
