+++ tutorials/postboard/client/src/routes/AppRoutes.tsx
@@ -1,19 +1,34 @@
-import React, { useContext } from 'react';
+import React, { useContext, FunctionComponent, SVGProps } from 'react';
 import Sidebar from 'components/Sidebar';
 import { AuthContext } from 'context/AuthContext';
-import useMenuItems from 'hooks/useMenuItems';
 import Login from 'views/Login';
 import { Route, Routes } from 'react-router-dom';
 import Home from 'pages/Home';
 import ViewPost from 'pages/ViewPost';
 import ExploreInput from 'components/ExploreInput';
+import { HomeSvg, ProfileSvg, ListSvg, BookmarkSvg, NotificationSvg, SettingsSvg } from '../assets/icons';
+
+export type MenuItem = {
+  iconComponent: FunctionComponent<SVGProps<SVGSVGElement>>;
+  label: string;
+  route: string;
+};
+
+const menuItems: Array<MenuItem> = [
+  { iconComponent: HomeSvg, label: 'Home', route: '/' },
+  { iconComponent: ProfileSvg, label: 'Profile', route: '/' },
+  { iconComponent: ListSvg, label: 'User list', route: '/' },
+  { iconComponent: BookmarkSvg, label: 'Bookmarks', route: '/' },
+  { iconComponent: NotificationSvg, label: 'Notifications', route: '/' },
+  { iconComponent: SettingsSvg, label: 'Settings', route: '/' },
+];
 
 const AppRoutes = () => {
   const authContext = useContext(AuthContext);
-  const items = useMenuItems({ isLoggedIn: !!authContext.state.address });
+  const isLoggedIn = !!authContext.state.address;
   return (
     <div className="container grid-container">
-      <Sidebar items={items} />
+      <Sidebar items={isLoggedIn ? [menuItems[0]] : menuItems} />
       <div className="app">
         <ExploreInput />
         <Routes>
